改写了省份排序算法，把控制台输出改为向文件输出，按作业要求整理代码文件
