+++ .gitignore
@@ -12,3 +12,4 @@
 InfectStatistic-example.iml
 /221701225/log/
 /221701225/result/
+/221701225/test/
